funcionando todo

+++ src/app/products/list/list.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { Product } from '../../models/product';
 import { ProductsService } from '../../services/products.service';
 import { ToastrService } from 'ngx-toastr';
+import { HttpErrorResponse } from '@angular/common/http';
 
 
 @Component({
@@ -33,19 +34,42 @@ export class ListComponent {
   }
 
   borrar(id: number) {
+    if (!id) {
+      console.error('Trying to delete with undefined ID');
+      return;
+    }
     this.productService.delete(id).subscribe(
-      data => {
+      () => {
         this.toastr.success('Producto Eliminado', 'OK', {
-          timeOut: 3000, positionClass: 'toast-top-center'
+          timeOut: 3000, 
+          positionClass: 'toast-top-center'
         });
         this.cargarProducts();
       },
-      err => {
-        this.toastr.error(err.error.mensaje, 'Fail', {
-          timeOut: 3000,  positionClass: 'toast-top-center',
+      (err: HttpErrorResponse) => { 
+        console.log(err);
+        let errorMessage = 'Error al intentar eliminar el producto.';
+        if (err.error instanceof ErrorEvent) {
+          // Error del lado del cliente
+          errorMessage = `Error: ${err.error.message}`;
+        } else if (err.status === 404) {
+          // Manejar específicamente el error 404
+          errorMessage = 'Producto no encontrado';
+        } else if (typeof err.error === 'string') {
+          // Manejar una respuesta de error que no es JSON
+          errorMessage = err.error;
+        } else if (err.error && err.error.mensaje) {
+          // Manejar una respuesta de error JSON
+          errorMessage = err.error.mensaje;
+        }
+        this.toastr.error(errorMessage, 'Fail', {
+          timeOut: 3000,  
+          positionClass: 'toast-top-center',
         });
       }
     );
   }
-
+  
+  
+  
 }
